src/pipeline/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared pipeline logic"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.0.0",
    "google-auth>=2.0.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "tenacity>=8.0.0",
    "typing-extensions>=4.0.0",
    "numpy>=1.24.4"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
